backend/requirements.txt
requests==2.25.1
gunicorn==20.0.4
Flask==1.1.2
flask-cors==3.0.10
Pillow==8.1.2
opencv-python-headless==4.5.1.48
ansicolors==1.1.8
rfc3339==6.2